@@ -1,6 +1,6 @@
 package br.com.bookstore.api.autor;
 
-import java.util.Date;
+import java.time.LocalDate;
 
 /**
  *
@@ -10,19 +10,19 @@ public class Autor {
     
     private int id;
     private String nome;
-    private Date dataNascimento;
+    private LocalDate dataNascimento;
     private Genero genero;   
 
     public Autor() {
     }
 
-    public Autor(String nome, Date dataNascimento, Genero genero) {
+    public Autor(String nome, LocalDate dataNascimento, Genero genero) {
         this.nome = nome;
         this.dataNascimento = dataNascimento;
         this.genero = genero;
     }
 
-    public Autor(int id, String nome, Date dataNascimento, Genero genero) {
+    public Autor(int id, String nome, LocalDate dataNascimento, Genero genero) {
         this.id = id;
         this.nome = nome;
         this.dataNascimento = dataNascimento;
@@ -46,11 +46,11 @@ public class Autor {
         this.nome = nome;
     }
 
-    public Date getDataNascimento() {
+    public LocalDate getDataNascimento() {
         return dataNascimento;
     }
 
-    public void setDataNascimento(Date dataNascimento) {
+    public void setDataNascimento(LocalDate dataNascimento) {
         this.dataNascimento = dataNascimento;
     }
 
